@@ -1,23 +1,19 @@
-import {ApplicationConfig, Component} from '@angular/core';
+import {ApplicationConfig, Component, importProvidersFrom} from '@angular/core';
 import {RouterOutlet} from '@angular/router';
-import {HttpClientModule} from "@angular/common/http";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 import {CommonModule} from "@angular/common";
 import {LoginComponent} from "./pages/login/login.component";
 import {ApiModule} from "./services/api.module";
 
 @Component({
-  selector: 'app-root',
-  standalone: true,
-  imports: [
-    RouterOutlet,
-    HttpClientModule,
+  imports: [RouterOutlet,
     CommonModule,
     LoginComponent,
-    ApiModule.forRoot({rootUrl:'http://95.31.234.122:1111'}).ngModule
-
-  ],
-  templateUrl: './app.component.html',
-  styleUrl: './app.component.scss'
+    ApiModule.forRoot({rootUrl: 'http://localhost:9090/api/v1'}).ngModule],
+  providers: [],
+  selector: 'app-root',
+  standalone: true, styleUrl: './app.component.scss',
+  templateUrl: './app.component.html'
 })
 export class AppComponent {
   title = 'book-network-ui';
