@@ -1,14 +1,14 @@
-import {HttpHeaders, HttpInterceptorFn} from '@angular/common/http';
+import { HttpHeaders, HttpInterceptorFn } from '@angular/common/http';
 import {inject} from "@angular/core";
-import {TokenService} from "../token/token.service";
+import {KeycloakService} from "../keycloak/keycloak.service";
 
 export const httpTokenInterceptor: HttpInterceptorFn = (req, next) => {
-  const tokenService = inject(TokenService);
-  const token = tokenService.token;
+  const keycloakService = inject(KeycloakService);
+  const token = keycloakService.keycloak?.token;
   if(token) {
     const authReq = req.clone({
       setHeaders: {
-        Authorization: 'Bearer ' + token
+        Authorization: `Bearer ${token}`
       }
     });
     return next(authReq);
