@@ -25,6 +25,7 @@ public class BookMapper { // Класс, ответственный за пре�
         return Book.builder() // Используем "строитель", чтобы легко создать объект Book
                 .id(request.id()) // Устанавливает идентификатор книги (может быть null для новых объектов)
                 .title(request.title()) // Устанавливает название книги
+                .isbn(request.isbn())
                 .authorName(request.authorName()) // Устанавливает имя автора
                 .synopsis(request.synopsis()) // Устанавливает краткое описание книги
                 .archived(false) // По умолчанию книга не архивирована
@@ -42,7 +43,7 @@ public class BookMapper { // Класс, ответственный за пре�
                 .rate(book.getRate()) // Устанавливаем рейтинг книги
                 .archived(book.isArchived()) // Устанавливаем флаг, указывающий, что книга архивирована
                 .shareable(book.isShareable()) // Устанавливаем флаг, указывающий, что книгу можно делить или передавать
-                .owner(book.getOwner().getFullName()) // Устанавливаем полное имя владельца книги
+//                .owner(book.getOwner().getFullName()) // Устанавливаем полное имя владельца книги
                 .cover(FileUtils.readFileFromLocation(book.getBookCover())) // Читаем обложку книги с помощью FileUtils и устанавливаем
                 .build(); // Завершаем создание объекта BookResponse и возвращаем его
     }
