@@ -13,7 +13,7 @@ public interface BookRepository extends JpaRepository<Book, Integer>, JpaSpecifi
             FROM Book book 
             WHERE book.archived = false 
             and book.shareable = true 
-            and book.owner.id != :userId 
+            and book.createdBy != :userId 
             """)
-    Page<Book> findAllDisplayableBooks(Pageable pageable, Integer userId);
+    Page<Book> findAllDisplayableBooks(Pageable pageable, String userId);
 }
