@@ -46,13 +46,13 @@ public class BookService {
      */
     public Integer save(BookRequest request, Authentication connectedUser) {
         // Получение информации о текущем пользователе из аутентификации
-        User user = ((User) connectedUser.getPrincipal());
+//        User user = ((User) connectedUser.getPrincipal());
 
         // Преобразование данных запроса в объект Book с помощью маппера
         Book book = bookMapper.toBook(request);
 
         // Установка владельца книги на текущего пользователя
-        book.setOwner(user);
+//        book.setOwner(user);
 
         // Сохранение книги в репозитории и возврат идентификатора сохраненной книги
         return bookRepository.save(book).getId();
@@ -83,13 +83,13 @@ public class BookService {
      */
     public PageResponse<BookResponse> findAllBooks(int page, int size, Authentication connectedUser) {
         // Получение объекта пользователя из текущей аутентификации
-        User user = ((User) connectedUser.getPrincipal());
+//        User user = ((User) connectedUser.getPrincipal());
 
         // Определение параметров постраничного запроса: номер страницы, размер страницы и сортировка по дате создания по убыванию
         Pageable pageable = PageRequest.of(page, size, Sort.by("createdDate").descending());
 
         // Выполнение запроса для получения всех отображаемых книг с учетом постраничного запроса и идентификатора пользователя
-        Page<Book> books = bookRepository.findAllDisplayableBooks(pageable, user.getId());
+        Page<Book> books = bookRepository.findAllDisplayableBooks(pageable, connectedUser.getName());
 
         // Преобразование страницы книг в список объектов BookResponse с использованием метода map
         List<BookResponse> bookResponses = books.stream()
@@ -111,7 +111,7 @@ public class BookService {
     //    используется для поиска всех книг, принадлежащих определенному владельцу
     public PageResponse<BookResponse> findAllBooksByOwner(int page, int size, Authentication connectedUser) {
         // Получение объекта аутентификации и извлечение идентификатора владельца из объекта Principal
-        User user = ((User) connectedUser.getPrincipal());
+//        User user = ((User) connectedUser.getPrincipal());
 
         // Создание объекта Pageable для поддержки пагинации и сортировки
         Pageable pageable = PageRequest.of(
@@ -121,7 +121,7 @@ public class BookService {
         );
 
         // Выполнение запроса для поиска книг, принадлежащих владельцу с идентификатором user.getId()
-        Page<Book> books = bookRepository.findAll(BookSpecification.withOwnerId(user.getId()), pageable);
+        Page<Book> books = bookRepository.findAll(BookSpecification.withOwnerId(connectedUser.getName()), pageable);
 
         // Преобразование страниц книг в список объектов BookResponse
         List<BookResponse> bookResponses = books.stream() // превращаем список в поток
@@ -141,12 +141,12 @@ public class BookService {
     }
 
     public PageResponse<BorrowedBookResponse> findAllBorrowedBooks(int page, int size, Authentication connectedUser) {
-        User user = ((User) connectedUser.getPrincipal());
+//        User user = ((User) connectedUser.getPrincipal());
         Pageable pageable = PageRequest.of(
                 page, // номер страницы
                 size, // размер страницы
                 Sort.by("createdDate").descending()); // сортировка по убыванию даты создания
-        Page<BookTransactionHistory> allBorrowedBooks = bookTransactionHistoryRepository.finAllBorrowedBooks(pageable, user.getId());
+        Page<BookTransactionHistory> allBorrowedBooks = bookTransactionHistoryRepository.finAllBorrowedBooks(pageable, connectedUser.getName());
         List<BorrowedBookResponse> bookResponses = allBorrowedBooks.stream()
                 .map(bookMapper::toBorrowedBookResponse)
                 .toList();
@@ -162,12 +162,12 @@ public class BookService {
     }
 
     public PageResponse<BorrowedBookResponse> findAllReturnedBooks(int page, int size, Authentication connectedUser) {
-        User user = ((User) connectedUser.getPrincipal());
+//        User user = ((User) connectedUser.getPrincipal());
         Pageable pageable = PageRequest.of(
                 page, // номер страницы
                 size, // размер страницы
                 Sort.by("createdDate").descending()); // сортировка по убыванию даты создания
-        Page<BookTransactionHistory> allBorrowedBooks = bookTransactionHistoryRepository.finAllReturnedBooks(pageable, user.getId());
+        Page<BookTransactionHistory> allBorrowedBooks = bookTransactionHistoryRepository.finAllReturnedBooks(pageable, connectedUser.getName());
         List<BorrowedBookResponse> bookResponses = allBorrowedBooks.stream()
                 .map(bookMapper::toBorrowedBookResponse)
                 .toList();
@@ -185,8 +185,8 @@ public class BookService {
     public Integer updateShareableStatus(Integer bookId, Authentication connectedUser) {
         Book book = bookRepository.findById(bookId)
                 .orElseThrow(() -> new EntityNotFoundException("No book found with the ID:: " + bookId));
-        User user = ((User) connectedUser.getPrincipal());
-        if (!Objects.equals(book.getOwner().getId(), user.getId())) {
+//        User user = ((User) connectedUser.getPrincipal());
+        if (!Objects.equals(book.getCreatedBy(), connectedUser.getName())) {
             throw new OperationNotPermittedException("You can't update others books shareable status");
         }
         book.setShareable(!book.isShareable());
@@ -197,8 +197,8 @@ public class BookService {
     public Integer updateArchivedStatus(Integer bookId, Authentication connectedUser) {
         Book book = bookRepository.findById(bookId)
                 .orElseThrow(() -> new EntityNotFoundException("No book found with the ID:: " + bookId));
-        User user = ((User) connectedUser.getPrincipal());
-        if (!Objects.equals(book.getOwner().getId(), user.getId())) {
+//        User user = ((User) connectedUser.getPrincipal());
+        if (!Objects.equals(book.getCreatedBy(),connectedUser.getName())) {
             throw new OperationNotPermittedException("You can't update others books archived status");
         }
         book.setArchived(!book.isArchived());
@@ -213,16 +213,16 @@ public class BookService {
         if (book.isArchived() || !book.isShareable()) {
             throw new OperationNotPermittedException("The requested book cannot be borrowed since it is archived or not shareable");
         }
-        User user = ((User) connectedUser.getPrincipal());
-        if (Objects.equals(book.getOwner().getId(), user.getId())) {
+//        User user = ((User) connectedUser.getPrincipal());
+        if (Objects.equals(book.getCreatedBy(), connectedUser.getName())) {
             throw new OperationNotPermittedException("You can't borrow your own book");
         }
-        final boolean isAlreadyBorrowed = bookTransactionHistoryRepository.isAlreadyBorrowedByUser(bookId, user.getId());
+        final boolean isAlreadyBorrowed = bookTransactionHistoryRepository.isAlreadyBorrowedByUser(bookId, connectedUser.getName());
         if (isAlreadyBorrowed) {
             throw new OperationNotPermittedException("The requested book is already borrowed");
         }
         BookTransactionHistory bookTransactionHistory = BookTransactionHistory.builder()
-                .user(user)
+                .userId(connectedUser.getName())
                 .book(book)
                 .returnApproved(false)
                 .build();
@@ -235,11 +235,11 @@ public class BookService {
         if (book.isArchived() || !book.isShareable()) {
             throw new OperationNotPermittedException("The requested book cannot be borrowed since it is archived or not shareable");
         }
-        User user = ((User) connectedUser.getPrincipal());
-        if (Objects.equals(book.getOwner().getId(), user.getId())) {
+//        User user = ((User) connectedUser.getPrincipal());
+        if (Objects.equals(book.getCreatedBy(), connectedUser.getName())) {
             throw new OperationNotPermittedException("You can't borrow or return your own book");
         }
-        BookTransactionHistory bookTransactionHistory = bookTransactionHistoryRepository.findByBookIdAndUserId(bookId, user.getId())
+        BookTransactionHistory bookTransactionHistory = bookTransactionHistoryRepository.findByBookIdAndUserId(bookId, connectedUser.getName())
                 .orElseThrow(() -> new EntityNotFoundException("You did not borrow this book"));
         bookTransactionHistory.setReturned(true);
         return bookTransactionHistoryRepository.save(bookTransactionHistory).getId();
@@ -251,11 +251,11 @@ public class BookService {
         if (book.isArchived() || !book.isShareable()) {
             throw new OperationNotPermittedException("The requested book cannot be borrowed since it is archived or not shareable");
         }
-        User user = ((User) connectedUser.getPrincipal());
-        if (!Objects.equals(book.getOwner().getId(), user.getId())) {
+//        User user = ((User) connectedUser.getPrincipal());
+        if (!Objects.equals(book.getCreatedBy(), connectedUser.getName())) {
             throw new OperationNotPermittedException("You can't return a book that you do not own");
         }
-        BookTransactionHistory bookTransactionHistory = bookTransactionHistoryRepository.findByBookIdAndOwnerId(bookId, user.getId())
+        BookTransactionHistory bookTransactionHistory = bookTransactionHistoryRepository.findByBookIdAndOwnerId(bookId, connectedUser.getName())
                 .orElseThrow(() -> new OperationNotPermittedException("The book is not returned yet. You cannot approve its return"));
         bookTransactionHistory.setReturnApproved(true);
         return bookTransactionHistoryRepository.save(bookTransactionHistory).getId();
@@ -275,10 +275,10 @@ public class BookService {
                 .orElseThrow(() -> new EntityNotFoundException("No book found with the ID:: " + bookId)); // Если книга не найдена, выбрасывается исключение
 
         // Получение идентификации аутентифицированного пользователя
-        User user = ((User) connectedUser.getPrincipal());
+//        User user = ((User) connectedUser.getPrincipal());
 
         // Сохранение файла обложки в файловой системе или облачном хранилище
-        var bookCover = fileStorageService.saveFile(file, user.getId()); // Сохраняет файл и возвращает путь к сохраненному файлу
+        var bookCover = fileStorageService.saveFile(file, connectedUser.getName()); // Сохраняет файл и возвращает путь к сохраненному файлу
 
         // Установка сохраненного пути обложки в объект книги
         book.setBookCover(bookCover); // Обновляет обложку в сущности книги
