@@ -9,7 +9,11 @@ import jakarta.persistence.Id;
 import lombok.*; // Lombok-аннотации для автоматического создания кода
 import lombok.experimental.FieldDefaults; // Позволяет установить уровень доступа к полям
 import lombok.experimental.SuperBuilder; // Позволяет использовать паттерн "строитель" для создания объектов
-import org.springframework.data.annotation.*; // Аннотации Spring Data для аудита
+
+import org.springframework.data.annotation.CreatedBy;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedBy;
+import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener; // Слушатель событий для аудита
 
 import java.time.LocalDateTime; // Для работы с локальными датами и временем
@@ -51,10 +55,10 @@ public class BaseEntity { // Базовый класс для других су�
     // Пользователь, который создал сущность
     @CreatedBy // Аннотация Spring Data, указывающая, что это поле хранит идентификатор создателя сущности
     @Column(nullable = false, updatable = false) // Поле не может быть null и не может быть обновлено
-            Integer createdBy; // Идентификатор пользователя, который создал сущность
+    String createdBy; // Идентификатор пользователя, который создал сущность
 
     // Пользователь, который последний раз изменил сущность
     @LastModifiedBy // Аннотация Spring Data, указывающая, что это поле хранит идентификатор последнего модификатора
     @Column(updatable = false) // Поле не может быть обновлено после создания сущности
-            Integer lastModifiedBy; // Идентификатор пользователя, который последний раз изменил сущность
+    String lastModifiedBy; // Идентификатор пользователя, который последний раз изменил сущность
 }
