@@ -3,6 +3,7 @@ package com.yaroslav.booknetwork.history;
 import com.yaroslav.booknetwork.book.Book;
 import com.yaroslav.booknetwork.common.BaseEntity;
 import com.yaroslav.booknetwork.user.User;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
@@ -19,9 +20,11 @@ import lombok.experimental.SuperBuilder;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class BookTransactionHistory extends BaseEntity {
 
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    User user;
+//    @ManyToOne
+//    @JoinColumn(name = "user_id")
+//    User user;
+    @Column(name = "user_id")
+    private String userId;
     @ManyToOne
     @JoinColumn(name = "book_id")
     Book book;
