@@ -5,6 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 import java.util.Optional;
@@ -14,44 +15,44 @@ public interface BookTransactionHistoryRepository extends JpaRepository<BookTran
     @Query("""
                         SELECT history
                         FROM BookTransactionHistory  history
-                        WHERE history.user.id = :userId
+                        WHERE history.userId = :userId
             """)
-    Page<BookTransactionHistory> finAllBorrowedBooks(Pageable pageable, Integer userId);
+    Page<BookTransactionHistory> finAllBorrowedBooks(Pageable pageable, String userId);
 
     @Query("""
                         SELECT history
                         FROM BookTransactionHistory  history
-                        WHERE history.book.owner.id = :userId
+                        WHERE history.book.createdBy = :userId
             """)
-    Page<BookTransactionHistory> finAllReturnedBooks(Pageable pageable, Integer userId);
+    Page<BookTransactionHistory> finAllReturnedBooks(Pageable pageable, String userId);
 
     @Query("""
                         SELECT 
                         (COUNT (*) > 0) AS isBorrowed
                         FROM BookTransactionHistory  bookTransactionHistory
-                        WHERE bookTransactionHistory.user.id = :userId
+                        WHERE bookTransactionHistory.userId = :userId
                         AND bookTransactionHistory.book.id = :bookId
                         AND bookTransactionHistory.returnApproved = false 
             """)
-    boolean isAlreadyBorrowedByUser(Integer bookId, Integer userId);
+    boolean isAlreadyBorrowedByUser(@Param("bookId")Integer bookId, @Param("userId") String userId);
 
     @Query("""
                         SELECT transaction
                         FROM BookTransactionHistory  transaction
-                        WHERE transaction.user.id = :userId
+                        WHERE transaction.userId = :userId
                         AND transaction.book.id = :bookId
                         AND transaction.returned = false 
                         AND transaction.returnApproved = false 
             """)
-    Optional<BookTransactionHistory> findByBookIdAndUserId(Integer bookId, Integer userId);
+    Optional<BookTransactionHistory> findByBookIdAndUserId(@Param("bookId")Integer bookId, @Param("userId") String userId);
 
     @Query("""
                         SELECT transaction
                         FROM BookTransactionHistory  transaction
-                        WHERE transaction.book.owner.id = :userId
+                        WHERE transaction.book.createdBy = :userId
                         AND transaction.book.id = :bookId
                         AND transaction.returned = true 
                         AND transaction.returnApproved = false 
             """)
-    Optional<BookTransactionHistory> findByBookIdAndOwnerId(Integer bookId, Integer userId);
+    Optional<BookTransactionHistory> findByBookIdAndOwnerId(@Param("bookId")Integer bookId, @Param("userId") String userId);
 }
