@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RoleRepository extends JpaRepository<Role,Integer> {
+public interface RoleRepository {//extends JpaRepository<Role,Integer> {
 
-    Optional<Role> findByName(String role);
+//    Optional<Role> findByName(String role);
 }
