@@ -19,9 +19,9 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-@Service
+//@Service
 @RequiredArgsConstructor
-public class UserDetailsServiceImpl implements UserDetailsService {
+public class UserDetailsServiceImpl {//implements UserDetailsService {
 
     private final UserRepository repository;
 
@@ -29,10 +29,10 @@ public class UserDetailsServiceImpl implements UserDetailsService {
      * Поиск пользователя по email и загрузка его деталей для аутентификации.
      * Если пользователь не найден, выбрасывается исключение UsernameNotFoundException.
      */
-    @Override
-    @Transactional
-    public UserDetails loadUserByUsername(String userEmail) throws UsernameNotFoundException {
-        return repository.findByEmail(userEmail)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-    }
+//    @Override
+//    @Transactional
+//    public UserDetails loadUserByUsername(String userEmail) throws UsernameNotFoundException {
+//        return repository.findByEmail(userEmail)
+//                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+//    }
 }
