@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository {//extends JpaRepository<Token, Integer> {
 
-    Optional<Token> findByToken(String token);
+    //Optional<Token> findByToken(String token);
 }
