@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository {//extends JpaRepository<User, Integer> {
 
-    Optional<User> findByEmail(String email);
+//    Optional<User> findByEmail(String email);
 }
